requirements.txt
# Core dependencies
streamlit>=1.40.0
python-dotenv>=1.0.0
pillow>=10.0.0

# Google Gemini SDK
google-genai>=1.0.0

# Install this package in editable mode
-e .
